[ui] Event flags dialog: fix crash and move to own file
