+++ src/emulator/gui/src/ui_mutexes_dialog.cpp
@@ -59,20 +59,3 @@ void DrawLwMutexesDialog(HostState &host) {
     }
     ImGui::End();
 }
-
-void DrawEventFlagsDialog(HostState &host) {
-    ImGui::Begin("Event Flags", &host.gui.eventflags_dialog);
-    ImGui::TextColored(ImVec4(255, 255, 0, 255), "%-16s %-32s   %-7s   %-8s  %-16s   %-16s", "ID", "EventFlag Name", "Flags", "Attributes", "Waiting Threads");
-
-    const std::lock_guard<std::mutex> lock(host.kernel.mutex);
-    for (auto event : host.kernel.eventflags) {
-        std::shared_ptr<EventFlag> event_state = event.second;
-        ImGui::Text("0x%08X       %-32s   %02d        %01d           %02u                 ",
-            event.first,
-            event_state->name,
-            event_state->flags,
-            event_state->attr,
-            event_state->waiting_threads.size());
-    }
-    ImGui::End();
-}
